add overlay to highlight selected line momentarily
Highlight the line for a second that was selected by the user in the cscope
buffer. Other emacs features (such as occur, compile, grep, etc) has this
functionality. Makes it easier to find the cursor when jumping around in the
code using cscope.

Set the cscope-allow-highlight-overlays variable in your .emacs if you want
this feature enabled.