add result filter functionality (C-c s R)
A global variable cscope-result-filter can be set via the interactive
cscope-set-result-filter function (or with the R key in the cscope buffer),
which is used to show only the results that contain the given string.

This can be very helpful, for example, if you are working on a large codebase
that is indexed entirely by cscope, and a variable appears in multiple modules
with the same name. If you are only interested in finding occurences in a
particular module, you can filter the cscope results to limit only to that module.

To reset the filter, just set it to empty string.