add function to list functions in current buffer using imenu
New interactive function cscope-find-function-list, which lists all the functions
from the current buffer in the cscope buffer, and makes them "selectable", as if
they were results of a lookup in cscope database. This helps navigating in large
source files with many functions.

The function uses imenu, so it works even if you don't have cscope database
prepared, which is handy if you just downloaded some code and want to look around
without involving cscope.